Finalização da atividade 9 de Estrutura de Dados

+++ 3o Sem/Estrutura/Ativ9/Juridica.java
@@ -1,16 +1,16 @@
 public class Juridica extends Pessoa {
     // Atributos
-    public int CNPJ;
+    public long CNPJ;
 
-    public Juridica(String nome, int idade, int CNPJ) {
+    public Juridica(String nome, int idade, long CNPJ) {
         super(nome, idade);
         this.CNPJ = CNPJ;
     }
     // Getters e setters
-    public void setCNPJ(int CNPJ) {
+    public void setCNPJ(long CNPJ) {
         this.CNPJ = CNPJ;
     }
-    public int getCNPJ() {
+    public long getCNPJ() {
         return this.CNPJ;
     }
